requirements.txt
# === Core Web Frameworks ===
fastapi
uvicorn[standard]
gunicorn

# === ML & Data Science ===
tensorflow-cpu>=2.11.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0
scipy>=1.7
keras>=2.11
joblib
h5py

# === Binance & Trading ===
python-binance
websockets
requests
APScheduler

# === FastAPI & API Utilities ===
python-multipart
python-dotenv
pydantic
httpx

# === Utility Libraries ===
python-dateutil
pytz
idna
certifi
charset-normalizer
importlib-metadata; python_version < "3.8"

# === OpenAI Integration ===
openai

# === Optional / Debug / Security ===
cryptography